refactor(capabilities): make input.rs an input/ submodule that owns its kinds
Closes #2164
