+++ crates/aether-capabilities/src/component/mod.rs
@@ -47,9 +47,11 @@ pub use route::{ComponentHostWasmExt, resolve_embedded};
 #[cfg(not(target_arch = "wasm32"))]
 mod load;
 
+#[cfg(not(target_arch = "wasm32"))]
+use crate::input::UnsubscribeAll;
 use aether_kinds::{DropComponent, ListComponents, LoadComponent, ReplaceComponent};
 #[cfg(not(target_arch = "wasm32"))]
-use aether_kinds::{LifecycleUnsubscribeAll, UnsubscribeAll};
+use aether_kinds::LifecycleUnsubscribeAll;
 
 #[cfg(not(target_arch = "wasm32"))]
 pub use native::ComponentHostConfig;
