+++ crates/aether-kinds/src/lib.rs
@@ -348,7 +348,7 @@ pub struct LifecycleUnsubscribeSelf {
 /// lifecycle stage's subscriber set. Issued by
 /// `ComponentHostCapability` on `DropComponent` so the lifecycle cap's
 /// per-stage broadcast doesn't keep firing at a dropped trampoline —
-/// the lifecycle-family counterpart of [`UnsubscribeAll`] for
+/// the lifecycle-family counterpart of `UnsubscribeAll` for
 /// `aether.input`. Idempotent: a mailbox with no stage subscriptions
 /// is still a no-op. Fire-and-forget; no reply. Cast-shape (Pod), one
 /// `mailbox` field, matching the sibling lifecycle kinds' raw-`u64`
@@ -374,7 +374,7 @@ pub struct LifecycleUnsubscribeAll {
 /// Reply to [`LifecycleSubscribe`] / [`LifecycleUnsubscribe`].
 /// `Err` carries the stage kind id and a human-readable reason —
 /// fail-fast subscribe per ADR-0082 §7. Same shape and rationale as
-/// [`SubscribeInputResult`] for input subscriptions.
+/// `SubscribeInputResult` for input subscriptions.
 #[derive(
     aether_data::Kind, aether_data::Schema, serde::Serialize, serde::Deserialize, Debug, Clone,
 )]
@@ -1340,99 +1340,6 @@ mod control_plane {
     #[kind(name = "aether.component.list_result")]
     pub struct ListComponentsResult {
         pub names: Vec<String>,
-    }
-
-    // ADR-0021 publish/subscribe routing for substrate input streams,
-    // ADR-0068 keying. The substrate maintains one subscriber set per
-    // input `KindId`; a `SubscribeInput` names the kind id and the
-    // mailbox to add. Issue 638 Phase 2 rehomed these kinds from
-    // `aether.control.*` to `aether.input.*`; the chassis-owned
-    // `InputCapability` handles them inline and replies via
-    // reply-to-sender.
-
-    /// `aether.input.subscribe` — add `mailbox` to the subscriber set
-    /// for `kind`. Idempotent: subscribing a mailbox already in the
-    /// set is still `Ok` (subscriptions are a set, not a counter).
-    /// Reply: `SubscribeInputResult`.
-    #[derive(aether_data::Kind, aether_data::Schema, Serialize, Deserialize, Debug, Clone)]
-    #[kind(name = "aether.input.subscribe")]
-    pub struct SubscribeInput {
-        pub kind: aether_data::KindId,
-        pub mailbox: aether_data::MailboxId,
-    }
-
-    /// `aether.input.subscribe_self` — reflexive counterpart of
-    /// [`SubscribeInput`]: subscribe the *sending* actor to the input
-    /// stream for `kind`, with no explicit `mailbox` field. The cap
-    /// resolves the subscriber from the inbound envelope's host-stamped
-    /// `Source` (ADR-0083) via `ctx.source_mailbox()`, so the
-    /// subscriber cannot be forged and the op is gated to in-process
-    /// actors by construction — an external session or another engine
-    /// has no local mailbox and gets an `Err` reply, pushing it onto
-    /// the named [`SubscribeInput`] form. This is the common
-    /// "subscribe me" case. Reply: `SubscribeInputResult`.
-    #[derive(aether_data::Kind, aether_data::Schema, Serialize, Deserialize, Debug, Clone)]
-    #[kind(name = "aether.input.subscribe_self")]
-    pub struct SubscribeInputSelf {
-        pub kind: aether_data::KindId,
-    }
-
-    /// `aether.input.unsubscribe` — remove `mailbox` from the
-    /// subscriber set for `kind`. Idempotent: unsubscribing a mailbox
-    /// that isn't subscribed is still `Ok`. Reply:
-    /// `SubscribeInputResult`.
-    #[derive(aether_data::Kind, aether_data::Schema, Serialize, Deserialize, Debug, Clone)]
-    #[kind(name = "aether.input.unsubscribe")]
-    pub struct UnsubscribeInput {
-        pub kind: aether_data::KindId,
-        pub mailbox: aether_data::MailboxId,
-    }
-
-    /// `aether.input.unsubscribe_self` — reflexive counterpart of
-    /// [`UnsubscribeInput`]: unsubscribe the *sending* actor from the
-    /// input stream for `kind`, with no explicit `mailbox` field. The
-    /// cap resolves the subscriber from the inbound envelope's
-    /// host-stamped `Source` (ADR-0083), the same gating as
-    /// [`SubscribeInputSelf`]. Idempotent on "not currently
-    /// subscribed." Reply: `SubscribeInputResult`.
-    #[derive(aether_data::Kind, aether_data::Schema, Serialize, Deserialize, Debug, Clone)]
-    #[kind(name = "aether.input.unsubscribe_self")]
-    pub struct UnsubscribeInputSelf {
-        pub kind: aether_data::KindId,
-    }
-
-    /// Reply to subscribe / unsubscribe / `unsubscribe_all` (ADR-0021 §2).
-    /// Only failure mode: the target mailbox id doesn't name a live
-    /// component (unknown, a sink, or already dropped).
-    #[derive(aether_data::Kind, aether_data::Schema, Serialize, Deserialize, Debug, Clone)]
-    #[kind(name = "aether.input.subscribe_result")]
-    pub enum SubscribeInputResult {
-        Ok,
-        Err { error: String },
-    }
-
-    /// `aether.input.unsubscribe_all` — remove `mailbox` from every
-    /// input stream's subscriber set. Issued by
-    /// `ComponentHostCapability` on `DropComponent` so the cap's
-    /// fan-out tables don't keep firing at a dropped trampoline.
-    /// Idempotent: a mailbox with no subscriptions is still a no-op.
-    /// Fire-and-forget; no reply. Cast-shape (Pod) — one
-    /// `MailboxId`, fixed size.
-    #[repr(C)]
-    #[derive(
-        Copy,
-        Clone,
-        Debug,
-        PartialEq,
-        Eq,
-        bytemuck::Pod,
-        bytemuck::Zeroable,
-        aether_data::Kind,
-        aether_data::Schema,
-    )]
-    #[kind(name = "aether.input.unsubscribe_all")]
-    pub struct UnsubscribeAll {
-        pub mailbox: aether_data::MailboxId,
     }
 
     /// Reference-image comparison for a `CaptureFrame.similarity` request
@@ -3564,11 +3471,6 @@ mod tests {
         assert_eq!(LoadResult::NAME, "aether.component.load_result");
         assert_eq!(DropResult::NAME, "aether.component.drop_result");
         assert_eq!(ReplaceResult::NAME, "aether.component.replace_result");
-        assert_eq!(SubscribeInput::NAME, "aether.input.subscribe");
-        assert_eq!(SubscribeInputSelf::NAME, "aether.input.subscribe_self");
-        assert_eq!(UnsubscribeInput::NAME, "aether.input.unsubscribe");
-        assert_eq!(UnsubscribeInputSelf::NAME, "aether.input.unsubscribe_self");
-        assert_eq!(SubscribeInputResult::NAME, "aether.input.subscribe_result");
         assert_eq!(CaptureFrame::NAME, "aether.render.capture_frame");
         assert_eq!(
             CaptureFrameResult::NAME,
